requirements.txt
# ALL COMMENTED PACKAGES ARE ONLY NECESSARY TO CONSTRUCT DATABASE #

beautifulsoup4==4.12.2
#faiss-cpu==1.7.4
fitz==0.0.1.dev2
numpy==1.24.3
openai==1.3.6
pymongo==3.12.0
python-dotenv==1.0.0
Requests==2.31.0
#scipy==1.10.0
#torch==1.11.0              # I don't even see this being used #
#transformers==4.32.1
Flask==2.2.5
Flask-Cors==3.0.10